requirements.txt
PySide6==6.8.1
requests==2.31.0
python-dotenv==1.0.0 